Fallback to session storage
